pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vault-reconstruct"
version = "0.1.0"
description = "Obsidian vault reconstruction + improvement toolkit"
requires-python = ">=3.10"
dependencies = [
    "tqdm",
    "python-dotenv",
    "ollama",
    "google-genai",
    "openai",
    "genanki",
    "textual",
    "sentence-transformers>=5.4.1",
    "arxiv>=3.0.0",
    "wikipedia>=1.4.0",
    "biopython>=1.87",
    "torch==2.6.0+cu124",
    "torchvision==0.21.0+cu124",
    "torchaudio==2.6.0+cu124",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[project.scripts]
vault-recon = "vault_hud:main_entry"

[tool.hatch.build.targets.wheel]
# Include the organizational directories and the core package.
include = [
    "vault_reconstruct/",
    "tools/",
    "scripts/",
    "vault_hud.py",
]
